requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
PyPDF2==3.0.1
nltk==3.8.1
sentence-transformers==2.7.0
chromadb==0.4.15
numpy==1.26.0
requests==2.31.0
pydantic==2.4.2
huggingface_hub==0.23.0
python-dotenv==1.0.1